pyproject.toml
[project]
name = "ai-expenses-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "anthropic>=0.73.0",
    "apscheduler>=3.11.1",
    "httpx>=0.28.1",
    "openai>=2.8.0",
    "docling>=1.0.0",
    "pillow>=10.0.0, <12.0.0",
    "psycopg2-binary>=2.9.11",
    "python-telegram-bot>=22.5",
    "sqlmodel>=0.0.27",
    "python-decouple>=3.8",
    "pydantic>=2.0.0",
]

[dependency-groups]
dev = [
    "ipykernel>=7.1.0",
]
